src/connector/testing/requirements.txt
livekit
livekit-api
fastapi[standard]
websockets
python-dotenv
opencv-python